Furnace and I broke stuff
I seem to have broken the movable_item script. All of a sudden it stopped working and then it didnt want to work again. I also have no fucking clue as to why the items dont want to use the correct bounds even though littearly everiything points to the fact that they should.
- Also I the furnace now makes a ding :D